fix detection of CVE-2015-6616 patches There seemed to be one wrong bug number in CVE_2015_6616.java.
Instead of bug 24630158 (of CVE-2015-6616), bug 24445127 (from CVE-2015-6620)
was checked. The list of bugs in the comments had a wrong URL for bug 24630158 too.

The detection of 24630158 is done in libstagefright_soft_mpeg4dec.so,
the other 2 bugs that were part of this check are still
detected in libstagefright.so